+++ CHANGELOG.md
@@ -6,13 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Nothing yet
+- WIP: Create a new preset directly from SelectablesListenerManager inspector
 
 ### Added
-- First version
+- Add Presets
 
 ### Changed
-- Nothing yet
+- Rename GlobalUiEventsManager -> SelectablesListenerManager
 
 ### Deprecated
 - Nothing yet
@@ -22,6 +22,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Nothing yet
+
+## [0.1.0] - 2025-10-07
+
+### Added
+- Add Presets
+
+### Changed
+- Rename GlobalUiEventsManager -> SelectablesListenerManager
+- Rename component and context menus
 
 ## [0.0.1] - 2025-09-29